+++ src/steps/vulnerabilities.ts
@@ -1,10 +1,9 @@
 import {
   createDirectRelationship,
   createIntegrationEntity,
-  // Entity,
+  Entity,
   IntegrationStep,
   IntegrationStepExecutionContext,
-  // JobState,
   RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
 
@@ -45,46 +44,60 @@ export function getVulnerabilityKey(id: string): string {
   return `insightvm_vulnerability:${id}`;
 }
 
-// async function findOrCreateVulnerability(
-//   jobState: JobState,
-//   assetVulnerability: InsightVmAssetVulnerability,
-// ): Promise<Entity> {
-//   const existingVulnerability = await jobState.findEntity(
-//     getVulnerabilityKey(assetVulnerability.id),
-//   );
+async function findOrCreateVulnerability(
+  { logger, jobState }: IntegrationStepExecutionContext<IntegrationConfig>,
+  assetVulnerability: InsightVmAssetVulnerability,
+): Promise<Entity> {
+  const existingVulnerability = await jobState.findEntity(
+    getVulnerabilityKey(assetVulnerability.id),
+  );
 
-//   if (existingVulnerability) {
-//     return existingVulnerability;
-//   }
+  if (existingVulnerability) {
+    logger.debug(
+      {
+        id: assetVulnerability.id,
+      },
+      'Found existing vulnerability entity in jobState.',
+    );
+    return existingVulnerability;
+  }
+  logger.debug(
+    {
+      id: assetVulnerability.id,
+    },
+    'Creating new vulnerability entity in jobState.',
+  );
+  return jobState.addEntity(createVulnerabilityEntity(assetVulnerability));
+}
 
-//   // TODO should fetch vulnerability from `/vulnerabilities/{id} endpoint to create entity.
-//   return jobState.addEntity(
-//     createIntegrationEntity({
-//       entityData: {
-//         source: assetVulnerability,
-//         assign: {
-//           _key: getVulnerabilityKey(assetVulnerability.id),
-//           _type: entities.VULNERABILITY._type,
-//           _class: entities.VULNERABILITY._class,
-//           id: `${assetVulnerability.id}`,
-//           name: assetVulnerability.id,
-//           category: 'other',
-//           severity: 'critical',
-//           blocking: false,
-//           open: false,
-//           production: false,
-//           public: true,
-//         },
-//       },
-//     }),
-//   );
-// }
+// TODO should fetch vulnerability from `/vulnerabilities/{id} endpoint to create entity.
+function createVulnerabilityEntity(
+  assetVulnerability: InsightVmAssetVulnerability,
+) {
+  return createIntegrationEntity({
+    entityData: {
+      source: assetVulnerability,
+      assign: {
+        _key: getVulnerabilityKey(assetVulnerability.id),
+        _type: entities.VULNERABILITY._type,
+        _class: entities.VULNERABILITY._class,
+        id: `${assetVulnerability.id}`,
+        name: assetVulnerability.id,
+        category: 'other',
+        severity: 'critical',
+        blocking: false,
+        open: false,
+        production: false,
+        public: true,
+      },
+    },
+  });
+}
 
-export async function fetchAssetVulnerabilities({
-  logger,
-  instance,
-  jobState,
-}: IntegrationStepExecutionContext<IntegrationConfig>) {
+export async function fetchAssetVulnerabilities(
+  context: IntegrationStepExecutionContext<IntegrationConfig>,
+) {
+  const { logger, instance, jobState } = context;
   const apiClient = createAPIClient(instance.config, logger);
 
   await jobState.iterateEntities(
@@ -113,18 +126,17 @@ export async function fetchAssetVulnerabilities({
             }),
           );
 
-          /* TODO: Re-implement. Temporarily disabled for testing. */
-          // const vulnerabilityEntity = await findOrCreateVulnerability(
-          //   jobState,
-          //   assetVulnerability,
-          // );
-          // await jobState.addRelationship(
-          //   createDirectRelationship({
-          //     _class: RelationshipClass.IS,
-          //     from: findingEntity,
-          //     to: vulnerabilityEntity,
-          //   }),
-          // );
+          const vulnerabilityEntity = await findOrCreateVulnerability(
+            context,
+            assetVulnerability,
+          );
+          await jobState.addRelationship(
+            createDirectRelationship({
+              _class: RelationshipClass.IS,
+              from: findingEntity,
+              to: vulnerabilityEntity,
+            }),
+          );
         },
       );
     },
@@ -153,3 +165,8 @@ export const vulnerabilitiesSteps: IntegrationStep<IntegrationConfig>[] = [
     executionHandler: fetchAssetVulnerabilities,
   },
 ];
+
+export const testFunctions = {
+  findOrCreateVulnerability,
+  createVulnerabilityEntity,
+};